+++ webpack.config.js
@@ -1,41 +1,62 @@
-var path = require('path');
-var webpack = require("webpack");
+const path = require('path');
+const webpack = require('webpack');
 
-module.exports = {
-    entry: [
-        "./src/Staction.ts"
-    ],
-    output: {
-        path: __dirname + '/build/',
-        filename: 'bundle.js',
-        publicPath: '/build/',
-        library: "staction",
-        libraryTarget: "umd",
-        umdNamedDefine: true,
-        globalObject: "typeof self !== 'undefined' ? self : this"
-    },
-    mode: 'production',
-    module: {
-      rules: [{
+const baseConfig = {
+  entry: ['./src/Staction.ts'],
+  output: {
+    path: __dirname + '/build/',
+    filename: 'bundle.js',
+    publicPath: '/build/',
+    library: 'staction',
+    libraryTarget: 'umd',
+    umdNamedDefine: true,
+  },
+  mode: 'production',
+  module: {
+    rules: [
+      {
         test: /\.ts$/,
         loader: 'babel-loader',
         include: path.join(__dirname, 'src'),
         options: {
-          presets: [
-            '@babel/preset-typescript',
-            '@babel/preset-env'
-          ],
+          presets: ['@babel/preset-typescript'],
           plugins: [
-            ["@babel/plugin-transform-runtime", {
-              "helpers": false,
-              "regenerator": true,
-            }],
             '@babel/plugin-proposal-object-rest-spread',
-            '@babel/plugin-proposal-class-properties'
-          ]
-        }
-      }]
-    },
-    resolve: { extensions: ['.js', '.jsx', '.tsx', '.ts', '.json'] },
-    devtool: 'source-map'
-}
+            '@babel/plugin-proposal-class-properties',
+          ],
+        },
+      },
+    ],
+  },
+  resolve: { extensions: ['.js', '.jsx', '.tsx', '.ts', '.json'] },
+  devtool: 'source-map',
+};
+
+const umdConfig = {
+  ...baseConfig,
+  output: {
+    filename: 'bundle.js',
+    path: __dirname + '/build/',
+    library: 'staction',
+    libraryTarget: 'umd',
+    publicPath: '/static/',
+    umdNamedDefine: true,
+  },
+};
+
+const esmConfig = {
+  ...baseConfig,
+  output: {
+    filename: 'esm-bundle.js',
+    path: __dirname + '/build/',
+    libraryTarget: 'module',
+    publicPath: '/static/',
+    umdNamedDefine: true,
+    module: true,
+  },
+  experiments: {
+    outputModule: true,
+  },
+};
+
+module.exports = [umdConfig, esmConfig];